Hoisted CopyrightFormatterResource to new directory & updated arch tests
